feat: add implementation for command split
