+++ cmd/kubectl-mongodb/migrate-to-mck/users.go
@@ -11,7 +11,6 @@ import (
 	"github.com/spf13/cobra"
 	k8svalidation "k8s.io/apimachinery/pkg/util/validation"
 
-	userv1 "github.com/mongodb/mongodb-kubernetes/api/v1/user"
 	"github.com/mongodb/mongodb-kubernetes/controllers/om"
 	"github.com/mongodb/mongodb-kubernetes/controllers/operator/authentication"
 	kubernetesClient "github.com/mongodb/mongodb-kubernetes/mongodb-community-operator/pkg/kube/client"
@@ -33,58 +32,16 @@ var uFlags usersFlags
 var UsersCmd = &cobra.Command{
 	Use:   "users",
 	Short: "Generate MongoDBUser Kubernetes CRs",
-	Long: `Generate MongoDBUser CRs from an Ops Manager or Cloud Manager automation config.
+	Long: `Generates MongoDBUser Kubernetes CRs from an Ops Manager/Cloud Manager automation
+config. The automation config is validated before generation. If any blockers are
+found, the command fails without producing output.
 
-The automation config is validated before output is produced. The command exits
-with an error if any blockers are found.
+Requires a ConfigMap (baseUrl, orgId, projectName) and a Secret (publicKey,
+privateKey) in the same format used by the operator.
 
-PREREQUISITES
+Example:
 
-  A ConfigMap and a Secret must exist in the target namespace before running this
-  command:
-
-    kubectl create configmap my-project \
-      --from-literal=baseUrl=<url> \
-      --from-literal=orgId=<id> \
-      --from-literal=projectName=<name>
-
-    kubectl create secret generic my-credentials \
-      --from-literal=publicKey=<key> \
-      --from-literal=privateKey=<key>
-
-  Each SCRAM user also requires a pre-created Secret containing their password
-  under the key "password":
-
-    kubectl create secret generic <secret-name> \
-      --from-literal=password=<password> \
-      -n <namespace>
-
-USAGE
-
-  Interactive mode: the command prompts for each user's Secret name in turn.
-  Press Enter to accept the suggested name (<username>-password).
-
-  Non-interactive mode: supply --users-secrets-file with a CSV file mapping
-  each user to a pre-created Secret:
-
-    # username:database,secret-name
-    alice:admin,alice-password
-    bob:reporting,bob-password
-
-EXAMPLES
-
-  Interactive:
-    kubectl mongodb migrate users \
-      --config-map-name my-project \
-      --secret-name my-credentials \
-      --namespace mongodb
-
-  Non-interactive:
-    kubectl mongodb migrate users \
-      --config-map-name my-project \
-      --secret-name my-credentials \
-      --namespace mongodb \
-      --users-secrets-file users-secrets.csv`,
+kubectl mongodb migrate users --config-map-name my-project --secret-name my-credentials --namespace mongodb`,
 	RunE: runGenerateUsers,
 }
 
@@ -93,7 +50,7 @@ func init() {
 	UsersCmd.Flags().StringVar(&uFlags.secretName, "secret-name", "", "Name of the Secret containing the OM API credentials (publicKey, privateKey) (required)")
 	UsersCmd.Flags().StringVar(&uFlags.namespace, "namespace", defaultNamespace, "Namespace of the ConfigMap and Secret")
 	UsersCmd.Flags().StringVarP(&uFlags.outputFile, "output", "o", "", "Write generated CRs to this file instead of stdout")
-	UsersCmd.Flags().StringVar(&uFlags.usersSecretsFile, "users-secrets-file", "", "CSV file mapping 'username:database,secret-name' for SCRAM users. Each line maps one user to a pre-created Kubernetes Secret. When omitted, the command prompts for each secret name interactively")
+	UsersCmd.Flags().StringVar(&uFlags.usersSecretsFile, "users-secrets-file", "", "CSV file mapping 'username:database,secret-name' for SCRAM users. When provided, customer-owned Secrets are referenced instead of generated and interactive prompts for user passwords are suppressed")
 	UsersCmd.Flags().StringVar(&uFlags.resourceNameOverride, "resource-name-override", "", "Name of the MongoDB replica set CR that the generated MongoDBUser resources will reference via mongodbResourceRef.name. Defaults to the normalized replica set name from the automation config")
 	_ = UsersCmd.MarkFlagRequired("config-map-name")
 	_ = UsersCmd.MarkFlagRequired("secret-name")
@@ -176,36 +133,23 @@ func collectUserCreds(ctx context.Context, kubeClient kubernetesClient.Client, a
 		}
 		return collectExistingUserSecrets(ctx, kubeClient, ac, opts, fileMapping)
 	}
-	return collectUserSecretNamesInteractively(ctx, kubeClient, ac, opts, scanner)
+	return collectUserPasswords(ac, opts, scanner)
 }
 
-func collectUserSecretNamesInteractively(ctx context.Context, kubeClient kubernetesClient.Client, ac *om.AutomationConfig, opts *GenerateOptions, scanner *bufio.Scanner) error {
-	users := scramUsers(ac)
-	if len(users) == 0 {
-		return nil
-	}
-
-	_, _ = fmt.Fprintf(promptOutput, "SCRAM users found. Create a Kubernetes Secret for each user before continuing:\n\n")
-	_, _ = fmt.Fprintf(promptOutput, "  kubectl create secret generic <secret-name> --from-literal=password=<password> -n %s\n\n", opts.Namespace)
-
-	opts.ExistingUserSecrets = make(map[string]string)
-	for _, user := range users {
-		suggestedName := userv1.NormalizeName(user.Username) + "-password"
-		input, err := promptLine(scanner, fmt.Sprintf("Secret name for user %q (db: %s) [%s]: ", user.Username, user.Database, suggestedName))
+func collectUserPasswords(ac *om.AutomationConfig, opts *GenerateOptions, scanner *bufio.Scanner) error {
+	opts.UserPasswords = make(map[string]string)
+	for _, user := range scramUsers(ac) {
+		password, err := promptLine(scanner, fmt.Sprintf("Enter password for SCRAM user %q (db: %s) [Enter to skip]: ", user.Username, user.Database))
 		if err != nil {
-			return fmt.Errorf("failed to read secret name for user %q: %w", user.Username, err)
+			return fmt.Errorf("failed to read password for user %q: %w", user.Username, err)
 		}
-		secretName := input
-		if secretName == "" {
-			secretName = suggestedName
+		if password == "" {
+			continue
 		}
-		if errs := k8svalidation.IsDNS1123Subdomain(secretName); len(errs) > 0 {
-			return fmt.Errorf("secret name %q for user %q is not a valid Kubernetes name: %s", secretName, user.Username, errs[0])
-		}
-		if err := validateUserSecret(ctx, kubeClient, user, secretName, ac, opts.Namespace); err != nil {
+		if err := validatePasswordAgainstOM(user.Username, user.Database, password, ac); err != nil {
 			return err
 		}
-		opts.ExistingUserSecrets[userKey(user.Username, user.Database)] = secretName
+		opts.UserPasswords[userKey(user.Username, user.Database)] = password
 	}
 	return nil
 }
